// src/config/enums.ts
export enum STATUS {
    PROCESSING = "in_progress",
    SCHEDULED = "scheduled",
    COMPLETED ="completed",
    FAILED ="failed"
}

export enum DIRECTION {
    INBOUND = "Inbound",
    OUTBOUND = "Outbound"
}

export enum YESNO {
    YES = "Yes",
    NO =  "No"
  }

  export enum FILE_TYPE  {
    AUDIO = 'audio',
    VIDEO = 'video',
    OTHER = 'txt'
  }

  export enum SENTIMENT  {
    POSITIVE = 'Positive',
    NEGATIVE = 'Negative',
    NUETRAL = 'Neutral'
  }